Removed an extra query that was unnecessary, changed EpisodeSQL to become one with Episode.java
