+++ src/main/java/org/example/Episode.java
@@ -1,17 +1,60 @@
 package org.example;
 
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "EpisodeList")
 public class Episode {
 
+    @Column(name="EpisodeID")
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int episodeID;
+    @Column(name="Universe")
+    private int universe;
+    @Column(name="Part")
+    private int part;
+    @Column(name="Number")
+    private int number;
+    @Column(name="Name")
     private String name;
-
+    @Column(name="Location")
+    private String location;
+    @Column(name="Latitude")
     private double latitude;
-
+    @Column(name="Longitude")
     private double longitude;
 
-    public Episode(String name, double latitude, double longitude) {
-        this.setName(name);
-        this.setLatitude(latitude);
-        this.setLongitude(longitude);
+    public int getEpisodeID() {
+        return episodeID;
+    }
+
+    public void setEpisodeID(int episodeID) {
+        this.episodeID = episodeID;
+    }
+
+    public int getUniverse() {
+        return universe;
+    }
+
+    public void setUniverse(int universe) {
+        this.universe = universe;
+    }
+
+    public int getPart() {
+        return part;
+    }
+
+    public void setPart(int part) {
+        this.part = part;
+    }
+
+    public int getNumber() {
+        return number;
+    }
+
+    public void setNumber(int number) {
+        this.number = number;
     }
 
     public String getName() {
@@ -19,8 +62,15 @@ public class Episode {
     }
 
     public void setName(String name) {
-
         this.name = name;
+    }
+
+    public String getLocation() {
+        return location;
+    }
+
+    public void setLocation(String location) {
+        this.location = location;
     }
 
     public double getLatitude() {
